Move fiels step to its own page
