+++ packages/component-submission/client/components/CoverLetterEditor.js
