+++ packages/component-submission/client/components/ManuscriptUpload.md
