+++ packages/component-submission/client/utils/ValidationSchemas.test.js
@@ -1,10 +1,10 @@
 import * as yup from 'yup'
 import { yupToFormErrors } from 'formik'
-import editorsPageSchema from './schema'
 
-const schema = yup.object().shape(editorsPageSchema)
+import { disclosureSchema, editorsSchema } from './ValidationSchemas'
 
-describe('Editors page form validation', () => {
+describe('Editors Schema', () => {
+  const schema = yup.object().shape(editorsSchema)
   it('allows valid data', () => {
     const validData = {
       suggestedSeniorEditors: [{ id: 1 }, { id: 2 }],
@@ -58,6 +58,51 @@ describe('Editors page form validation', () => {
       suggestedReviewingEditors: 'Please suggest at least 2 editors',
       suggestedSeniorEditors: 'Please suggest at least 2 editors',
       opposedReviewersReason: 'Please provide a reason for exclusion',
+    })
+  })
+})
+
+describe('Disclosure Schema', () => {
+  const schema = yup.object().shape(disclosureSchema)
+  it('allows valid data', () => {
+    const validData = {
+      submitterSignature: 'Jo Franchetti',
+      disclosureConsent: true,
+    }
+    expect(() => schema.validateSync(validData)).not.toThrow()
+  })
+
+  it('registers empty name as invalid', () => {
+    const invalidData = {
+      submitterSignature: '',
+      disclosureConsent: true,
+    }
+
+    let errors
+    try {
+      schema.validateSync(invalidData, { abortEarly: false })
+    } catch (e) {
+      errors = yupToFormErrors(e)
+    }
+
+    expect(errors).toEqual({ submitterSignature: 'Your name is required' })
+  })
+
+  it('registers an unchecked consent box as invalid', () => {
+    const invalidData = {
+      submitterSignature: 'Jo Franchetti',
+      disclosureConsent: false,
+    }
+
+    let errors
+    try {
+      schema.validateSync(invalidData, { abortEarly: false })
+    } catch (e) {
+      errors = yupToFormErrors(e)
+    }
+
+    expect(errors).toEqual({
+      disclosureConsent: 'We are unable to proceed without your consent',
     })
   })
 })
