+++ packages/component-submission/client/utils/stripHtml.js
@@ -1,0 +1,5 @@
+export default function stripHtml(html) {
+  const htmlWithLineBreaks = html.replace(/<\/p>/g, '</p>\n')
+  const doc = new DOMParser().parseFromString(htmlWithLineBreaks, 'text/html')
+  return doc.body.textContent || ''
+}
